backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
